Finished users basic api tests

+++ services/users/app/db/crud.py
@@ -117,7 +117,7 @@ def delete_user(session, email: str, password: str):
     return True
 
 
-def get_near_users(session, latitude: int, longitude: int, radius: int):
+def get_near_users(session, latitude: float, longitude: float, radius: int):
     """Get all users in range of radius (KM).
 
     Args:
@@ -129,6 +129,7 @@ def get_near_users(session, latitude: int, longitude: int, radius: int):
     Returns:
         list. User objects.
     """
+    # TODO: TEST!!!
     users_in_range = session.query(User).filter(func.acos(
         func.sin(func.radians(latitude)) * func.sin(
             func.radians(User.latitude)) + func.cos(
